palindrome-number solved in php

+++ c++/palindrome-number.cpp
@@ -4,7 +4,7 @@ using namespace std;
 class Solution {
 public:
     bool isPalindrome(int x) {
-        int rev_num = 0, initial = x;
+        long int rev_num = 0, initial = x;
 		while(x > 0) {
         	rev_num = rev_num * 10 + x%10;
         	x=x/10;
